requirements.txt
pygithub>=1.55,<1.56
click>=8.1.2,<8.2